[Feat] MapManager 생성, 기능 정리
