--- Assets/Prefabs/BackGround/Backgroundreverse.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 2cbcc761ad24cbd498d9bf3e4154e602
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
