+++ Assets/Prefabs/BackGround/Backgroundstraight.prefab
@@ -10,6 +10,7 @@ GameObject:
   m_Component:
   - component: {fileID: 9207805186883838742}
   - component: {fileID: 5053144983931544974}
+  - component: {fileID: 4484692700106473116}
   m_Layer: 0
   m_Name: Backgroundstraight
   m_TagString: Untagged
@@ -84,3 +85,16 @@ SpriteRenderer:
   m_WasSpriteAssigned: 1
   m_MaskInteraction: 0
   m_SpriteSortPoint: 0
+--- !u!114 &4484692700106473116
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 6015415691774125930}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 1d77c2e61f6075542b8f32e5e171367b, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  scrollSpeed: 30
