--- Assets/Scripts/Map/BackgroundManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b8cfbde0a88fa2944a06366c4e6ccc4d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
