--- Assets/Scripts/UI/PauseManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ff7cddae66552ef43b341a1245bf28d3
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
